+++ docs/CHANGE_LOG.md
@@ -2,6 +2,7 @@
 
 ## Version 2.0.5:
 - Added global Ctrl/Command + Shift + V support to paste the selected events (only for Full Day view).
+- Added a new setting "defaultEventDuration" (defaults to 30), which states the default duration a new event should use when being added.
 - 
 
 <br>